add tests and bug fixes

+++ src/interfaces/IERC7540Redeem.sol
@@ -20,7 +20,7 @@ $$/   $$/ $$/  $$$$$$$/ $$$$$$$/  $$/   $$/
  * @notice Terms of Service: https://www.othentic.xyz/terms-of-service
  */
 interface IERC7540Redeem {
-    function requestRedeem(uint256 _shares) external returns (uint40 _batchId);
+    function requestRedeem(uint256 _shares) external returns (uint48 _batchId);
 
-    function pendingRedeemRequest(uint40 _batchId) external view returns (uint256 _shares);
+    function pendingRedeemRequest(uint48 _batchId) external view returns (uint256 _shares);
 }
